refactor: re-export types, api, events, observation from looprs-core
